requirements.txt
async-generator==1.10
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
comtypes==1.1.10
configparser==5.2.0
crayons==0.4.0
cryptography==36.0.1
dominate==2.6.0
enum34==1.1.10
Flask==2.0.2
Flask-Bootstrap==3.3.7.1
future==0.18.2
h11==0.12.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
outcome==1.1.0
psutil==5.9.0
pycaw==20181226
pycparser==2.21
pyOpenSSL==21.0.0
python-dotenv==0.19.2
requests==2.26.0
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
visitor==0.1.3
webdriver-manager==3.5.2
Werkzeug==2.0.2
wsproto==1.0.0
